Refactor to improve readability
